chore(deps): update all javascript dependencies and adapt to typescript 6
Consolidates Dependabot PRs #122-#127: the 16-package minor/patch group,
typescript 6.0.3, @vitejs/plugin-react 6.0.2, tailwindcss 4.3.0,
tailwind-merge 3.6.0, and globals 17.6.0.

Compatibility fixes required by the new versions:
- packages/api: add explicit bun types to tsconfig (TypeScript 6 no
  longer picks them up automatically)
- packages/frontend: drop the deprecated baseUrl compiler option
- NetworkingPanel: move useMutation calls above the early returns
  (rules-of-hooks violations surfaced by eslint-plugin-react-hooks 7)
- SgRuleTable: extract non-component helper exports to sgRuleHelpers.ts
  (react-refresh/only-export-components)
