+++ packages/api/package.json
@@ -10,18 +10,18 @@
     "test": "bun test"
   },
   "dependencies": {
-    "@aws-sdk/client-ec2": "^3.1063.0",
-    "@aws-sdk/client-eks": "^3.1063.0",
-    "@aws-sdk/client-lambda": "^3.1063.0",
-    "@aws-sdk/client-rds": "^3.1063.0",
-    "@aws-sdk/client-s3": "^3.1063.0",
-    "@aws-sdk/client-secrets-manager": "^3.1063.0",
+    "@aws-sdk/client-ec2": "^3.1075.0",
+    "@aws-sdk/client-eks": "^3.1075.0",
+    "@aws-sdk/client-lambda": "^3.1075.0",
+    "@aws-sdk/client-rds": "^3.1075.0",
+    "@aws-sdk/client-s3": "^3.1075.0",
+    "@aws-sdk/client-secrets-manager": "^3.1075.0",
     "dotenv": "^17.4.2",
-    "hono": "^4.12.25"
+    "hono": "^4.12.27"
   },
   "devDependencies": {
     "@types/bun": "^1.3.2",
-    "@types/node": "^25.8.0",
-    "typescript": "^5.4.5"
+    "@types/node": "^25.9.4",
+    "typescript": "^6.0.3"
   }
 }
