+++ packages/frontend/package.json
@@ -12,33 +12,33 @@
     "type-check": "tsc --noEmit"
   },
   "dependencies": {
-    "@tanstack/react-query": "^5.101.0",
-    "@tanstack/react-query-devtools": "^5.101.0",
+    "@tanstack/react-query": "^5.101.2",
+    "@tanstack/react-query-devtools": "^5.101.2",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "date-fns": "^4.1.0",
-    "lucide-react": "^1.16.0",
+    "lucide-react": "^1.21.0",
     "react": "^19.2.6",
     "react-dom": "^19.2.0",
-    "react-router-dom": "^7.15.1",
-    "tailwind-merge": "^2.3.0",
+    "react-router-dom": "^7.18.0",
+    "tailwind-merge": "^3.6.0",
     "zustand": "^5.0.13"
   },
   "devDependencies": {
-    "@types/node": "^25.8.0",
+    "@types/node": "^25.9.4",
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.0",
-    "@vitejs/plugin-react": "^5.0.0",
-    "autoprefixer": "^10.4.19",
+    "@vitejs/plugin-react": "^6.0.2",
+    "autoprefixer": "^10.5.2",
     "eslint": "^10.4.0",
     "@eslint/js": "^10.0.1",
     "eslint-plugin-react-hooks": "^7.1.1",
     "eslint-plugin-react-refresh": "^0.5.2",
-    "globals": "^16.0.0",
-    "postcss": "^8.5.15",
-    "tailwindcss": "^3.4.3",
-    "typescript": "^5.4.5",
-    "typescript-eslint": "^8.60.1",
-    "vite": "^8.0.13"
+    "globals": "^17.6.0",
+    "postcss": "^8.5.16",
+    "tailwindcss": "^4.3.0",
+    "typescript": "^6.0.3",
+    "typescript-eslint": "^8.62.0",
+    "vite": "^8.1.0"
   }
 }
